Preserve calle-saved registers

+++ mc.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 NAME=""
+S_MODE=false
 
 get_file_name () {
     NAME=`basename $1`
@@ -13,12 +14,20 @@ if [ $# -eq 0 ] || [ $# -gt 2 ]; then
     printf "Usage: mc -printAst <filename>\n"
     printf "       mc -noRegAlloc <filename>\n"
     printf "       mc -dumpIL <filename>\n"
+    printf "       mc -S <filename>\n"
     exit 1
 fi
 
 if [[ ${1:0:1} == "-" ]]; then
-    COMPILER_ARG=$1
-    FPATH=$2
+    if [ $1 = "-S" ]; then
+        COMPILER_ARG=""
+        FPATH=$2
+        S_MODE=true
+    else
+        COMPILER_ARG=$1
+        FPATH=$2
+    fi
+
 else
     COMPILER_ARG=""
     FPATH=$1
@@ -41,10 +50,14 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-gcc "$NAME.s"
-
-# Remove temporary fies
-if [ $? -eq 0 ]; then
+if [ "$S_MODE" = true ]; then
     rm "$NAME.json"
-    rm "$NAME.s"
+else
+    gcc "$NAME.s"
+
+    # Remove temporary fies
+    if [ $? -eq 0 ]; then
+        rm "$NAME.json"
+        rm "$NAME.s"
+    fi
 fi
